Regen and copy to Antora structure
This regenerates the starter documentation, needed to add the
`:page-partial:` attribute at the top of each `.adoc` file and runs
copies the starter `.adoc` files under `docs/modules/ROOT/pages`.

We need the files in `docs/modules/ROOT/pages` in order for Antora to
find them and for the `include::` to function within the `components`
module in the main Camel git repository.

This could be replaced with symlinks (Antora dropped support for
symlinks) or if Antora would allow for different content organization.