fix: xref link for local components
We have the components from `core` module in the Camel catalog so when
generating the list of starters they get pulled in and linked to the
(main) Camel components reference where they do not exist.

Logic for detecting such components by looking if the `.adoc` files are
present looked at `component-starters` subdirectory of
`docs/modules/ROOT/pages` and did not detect them as local components.

This fixes that as all `.adoc` files are placed within
`docs/modules/ROOT/pages`.

Also cleans up the logic a bit.